+++ src/coreaudio_snd.rs
@@ -97,7 +97,7 @@ impl AudioContext {
 }
 
 pub struct Sound {
-    id: usize,
+    id: u32,
 }
 
 impl Sound {
